Move alert clearing to MessagePipelineProvider
