+++ packages/suite-base/src/components/MessagePipeline/index.tsx
@@ -20,6 +20,7 @@ import { StoreApi, useStore } from "zustand";
 import { useGuaranteedContext } from "@lichtblick/hooks";
 import { Immutable } from "@lichtblick/suite";
 import { AppSetting } from "@lichtblick/suite-base/AppSetting";
+import { useAlertsActions } from "@lichtblick/suite-base/context/AlertsContext";
 import CurrentLayoutContext, {
   LayoutState,
 } from "@lichtblick/suite-base/context/CurrentLayoutContext";
@@ -100,6 +101,8 @@ const selectSubscriptions = (state: MessagePipelineInternalState) => state.publi
 
 export function MessagePipelineProvider({ children, player }: ProviderProps): React.ReactElement {
   const promisesToWaitForRef = useRef<FramePromise[]>([]);
+  const previousPlayerRef = useRef<Player | undefined>();
+  const { clearAlerts } = useAlertsActions();
 
   // We make a new store when the player changes. This throws away any state from the previous store
   // and re-creates the pipeline functions and references. We make a new store to avoid holding onto
@@ -179,6 +182,14 @@ export function MessagePipelineProvider({ children, player }: ProviderProps): Re
     };
   }, [currentLayoutContext, player]);
 
+  // Session alerts from message converters should not carry over when switching to a new player.
+  useEffect(() => {
+    if (previousPlayerRef.current != undefined && previousPlayerRef.current !== player) {
+      clearAlerts();
+    }
+    previousPlayerRef.current = player;
+  }, [player, store, clearAlerts]);
+
   useEffect(() => {
     const dispatch = store.getState().dispatch;
     if (!player) {
@@ -196,6 +207,7 @@ export function MessagePipelineProvider({ children, player }: ProviderProps): Re
       msPerFrameRef,
       promisesToWaitForRef,
       store,
+      clearAlerts,
     });
     player.setListener(listener);
     return () => {
@@ -207,7 +219,7 @@ export function MessagePipelineProvider({ children, player }: ProviderProps): Re
         renderDone: undefined,
       });
     };
-  }, [player, store]);
+  }, [player, store, clearAlerts]);
 
   return <ContextInternal.Provider value={store}>{children}</ContextInternal.Provider>;
 }
@@ -251,11 +263,12 @@ function createPlayerListener(args: {
   msPerFrameRef: React.MutableRefObject<number>;
   promisesToWaitForRef: React.MutableRefObject<FramePromise[]>;
   store: StoreApi<MessagePipelineInternalState>;
+  clearAlerts: () => void;
 }): {
   listener: (state: PlayerState) => Promise<void>;
   cleanupListener: () => void;
 } {
-  const { msPerFrameRef, promisesToWaitForRef, store } = args;
+  const { msPerFrameRef, promisesToWaitForRef, store, clearAlerts } = args;
   const updateState = store.getState().dispatch;
   const messageOrderTracker = new MessageOrderTracker();
   let closed = false;
@@ -318,6 +331,7 @@ function createPlayerListener(args: {
     }
 
     if (prevPlayerId != undefined && listenerPlayerState.playerId !== prevPlayerId) {
+      clearAlerts();
       store.getState().reset();
     }
     prevPlayerId = listenerPlayerState.playerId;
